+++ algo3.cpp
@@ -69,7 +69,7 @@ int64 algo3(string filename, int64 n){
 
 
 int main(int argc, char *argv[]){
-    int n = atoi(argv[1]);
+    int64 n = atoll(argv[1]);
     clock_t t_start, t_end;
     srand(time(0));
     t_start = clock();
