+++ algo4.cpp
@@ -20,7 +20,7 @@ int algo4(string filename, int64 n){
 
 // main function, takes n as input (in command line)
 int main(int argc, char *argv[]){
-    int n = atoi(argv[1]);
+    int64 n = atoll(argv[1]);
     clock_t t_start, t_end;
     srand(time(0));
     t_start = clock();
